Organize Telegram bot project structure
