+++ app/__init__.py
@@ -1,0 +1,2 @@
+"""Core application package for the Telegram bot."""
+
