+++ app/handlers/__init__.py
@@ -1,0 +1,5 @@
+"""Handlers for the Telegram bot."""
+
+from . import start, topup, admin
+
+__all__ = ["start", "topup", "admin"]
