+++ app/utils/__init__.py
@@ -1,4 +1,4 @@
-from loader import bot, CHANNEL_ID
+from ..loader import bot, CHANNEL_ID
 
 BOT_NAME = ""
