+++ data/__init__.py
@@ -1,0 +1,2 @@
+"""Data storage for the Telegram bot."""
+
